criado controller e corrigido o loop

+++ TesteDesenvolvedor/TesteDesenvolvedor.Domain/PosicaoVeiculo.cs
@@ -10,7 +10,7 @@ namespace TesteDesenvolvedor.Domain
     {
         public double Latitude { get; set; }
         public double Longitude { get; set; }
-        public double VeiculoId { get; set; }
+        public long VeiculoId { get; set; }
         public Veiculo Veiculo { get; set; }
     }
 }
